adjsuted logic
